+++ data-agent-management/src/main/java/com/alibaba/cloud/ai/dataagent/service/datasource/AgentDatasourceService.java
@@ -16,31 +16,28 @@
 package com.alibaba.cloud.ai.dataagent.service.datasource;
 
 import com.alibaba.cloud.ai.dataagent.entity.AgentDatasource;
-
 import java.util.List;
 
 public interface AgentDatasourceService {
 
-	/**
-	 * Initialize agent's database schema using datasource
-	 */
+	/** Initialize agent's database schema using datasource */
 	Boolean initializeSchemaForAgentWithDatasource(Long agentId, Integer datasourceId, List<String> tables);
 
-	List<AgentDatasource> getAgentDatasource(Integer agentId);
+	List<AgentDatasource> getAgentDatasource(Long agentId);
 
-	default AgentDatasource getCurrentAgentDatasource(Integer agentId) {
+	default AgentDatasource getCurrentAgentDatasource(Long agentId) {
 		return getAgentDatasource(agentId).stream()
 			.filter(a -> a.getIsActive() != 0)
 			.findFirst()
 			.orElseThrow(() -> new IllegalStateException("Agent " + agentId + " has no active datasource"));
 	}
 
-	AgentDatasource addDatasourceToAgent(Integer agentId, Integer datasourceId);
+	AgentDatasource addDatasourceToAgent(Long agentId, Integer datasourceId);
 
-	void removeDatasourceFromAgent(Integer agentId, Integer datasourceId);
+	void removeDatasourceFromAgent(Long agentId, Integer datasourceId);
 
-	AgentDatasource toggleDatasourceForAgent(Integer agentId, Integer datasourceId, Boolean isActive);
+	AgentDatasource toggleDatasourceForAgent(Long agentId, Integer datasourceId, Boolean isActive);
 
-	void updateDatasourceTables(Integer agentId, Integer datasourceId, List<String> tables);
+	void updateDatasourceTables(Long agentId, Integer datasourceId, List<String> tables);
 
 }
